move labware outline into labware internals
