+++ components/src/hardware-sim/Labware/labwareInternals/LabwareOutline.tsx
