+++ components/src/hardware-sim/Labware/labwareInternals/index.ts
@@ -2,3 +2,4 @@ export * from './StaticLabware'
 export * from './StyledWells'
 export * from './WellLabels'
 export * from './FilledWells'
+export * from './LabwareOutline'
