+++ components/src/legacy-hardware-sim/index.ts
@@ -1,7 +1,5 @@
 export * from './LabwareNameOverlay'
-export * from './LabwareOutline'
 
 export * from './constants'
 export * from './LabwareNameOverlay'
-export * from './LabwareOutline'
 export * from './ModuleViz'
